% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myddt.R
\name{myddt}
\alias{myddt}
\title{myddt}
\usage{
myddt(df = ddt, SPECIES = "CCATFISH")
}
\arguments{
\item{df}{the DDT dataframe}

\item{SPECIES}{the species to subset df with}
}
\value{
Named list of dataframe before subsetting, dataframe after subsetting, and relative frequency of RIVER values before subsetting
}
\description{
For use with the DDT dataset from Statistics for Engineering and the Sciences, 6th. Subsets the DDT dataframe.
}
\examples{
myddt(ddt, "CCATFISH")
}
