% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mynbin.R
\name{mynbin}
\alias{mynbin}
\title{Title}
\usage{
mynbin(y, r, p)
}
\arguments{
\item{y}{Quantile}

\item{r}{Num success}

\item{p}{Probability of success}
}
\value{
Probabilty of falling within the lower tail of the quantile
}
\description{
Title
}
\examples{
\dontrun{mynbin(10, 3, 0.4)}
}
