% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myread.R
\name{myread}
\alias{myread}
\title{Title}
\usage{
myread(csv, dird)
}
\arguments{
\item{csv}{The name of a csv file}

\item{dird}{The directory the file is in}
}
\value{
The data frame inside the csv file
}
\description{
Title
}
\examples{
\dontrun{mpg.df=myread("EPAGAS.csv")}
}
